+++ test/Test/Notebook.hs
@@ -18,6 +18,7 @@ import ScriptHs.Notebook (
  )
 
 import ScriptHs.Parser (CabalMeta (metaDeps), Line (..))
+import ScriptHs.Run (defaultRunOptions)
 
 notebookTests :: TestTree
 notebookTests =
@@ -191,7 +192,7 @@ notebookTests =
             "processNotebook"
             [ testCase "no code blocks => returns input unchanged" $ do
                 let input = T.unlines ["# Title", "", "Just prose.", ""]
-                out <- processNotebook "" input
+                out <- processNotebook defaultRunOptions "" input
                 out @?= input
             , testCase "non-haskell code blocks only => returns input unchanged" $ do
                 let input =
@@ -204,7 +205,7 @@ notebookTests =
                             , ""
                             , "more prose"
                             ]
-                out <- processNotebook "" input
+                out <- processNotebook defaultRunOptions "" input
                 out @?= input
             ]
         ]